requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# LangChain & LangGraph
langchain>=0.1.10
langchain-groq>=0.1.0
langchain-community>=0.0.25
langgraph>=0.0.25

# Data Validation & Parsing
pydantic>=2.6.0
typing-extensions>=4.9.0

# Database
sqlalchemy>=2.0.0

# Environment Management
python-dotenv>=1.0.0

# Optional: For Streamlit UI (if needed)
# streamlit>=1.31.0
